+++ src/components/ImageLoader/BlurhashCanvas.test.jsx
@@ -123,7 +123,7 @@ describe('BlurhashCanvas', () => {
       expect(props.width).toBe('1440');
       expect(props.height).toBe('810');
       expect(props.data).toEqual(
-        '{"hash":"HASH","punch":1,"ratio":2,"width":32,"height":16}',
+        '{"hash":"HASH","punch":1,"ratio":2,"width":32,"height":16,"imgWidth":"1440px"}',
       );
     });
 
@@ -341,6 +341,52 @@ describe('BlurhashCanvas', () => {
         expect(canvas.children).toBe(null);
         expect(canvas.props.style.width).toBe('100%');
       });
+
+      test('accepts string as imgWidth', () => {
+        let component;
+        mockCanvas.offsetWidth = 100;
+        const mockPlaceholderExtraStyleRef = { current: {} };
+        act(() => {
+          component = create(
+            <BlurhashCanvas
+              hash="HASH"
+              ratio={2}
+              punch={1}
+              width={32}
+              imgWidth={'140'}
+              placeholderExtraStyleRef={mockPlaceholderExtraStyleRef}
+            />,
+            { createNodeMock: () => mockCanvas },
+          );
+        });
+        const canvas = component.toJSON();
+        expect(canvas.type).toBe('canvas');
+        expect(canvas.children).toBe(null);
+        expect(canvas.props.style.width).toBe('140px');
+      });
+
+      test('accepts pixels as imgWidth', () => {
+        let component;
+        mockCanvas.offsetWidth = 100;
+        const mockPlaceholderExtraStyleRef = { current: {} };
+        act(() => {
+          component = create(
+            <BlurhashCanvas
+              hash="HASH"
+              ratio={2}
+              punch={1}
+              width={32}
+              imgWidth={'140px'}
+              placeholderExtraStyleRef={mockPlaceholderExtraStyleRef}
+            />,
+            { createNodeMock: () => mockCanvas },
+          );
+        });
+        const canvas = component.toJSON();
+        expect(canvas.type).toBe('canvas');
+        expect(canvas.children).toBe(null);
+        expect(canvas.props.style.width).toBe('140px');
+      });
     });
 
     test('updates', () => {
@@ -367,6 +413,38 @@ describe('BlurhashCanvas', () => {
       expect(canvas.type).toBe('canvas');
       expect(canvas.children).toBe(null);
       expect(canvas.props.style.height).toBe(100);
+    });
+
+    test('does not update with zero offset width', () => {
+      let component;
+      mockCanvas.offsetWidth = 100;
+      const mockPlaceholderExtraStyleRef = { current: {} };
+      act(() => {
+        component = create(
+          <BlurhashCanvas
+            hash="HASH"
+            ratio={0.5}
+            punch={1}
+            width={32}
+            placeholderExtraStyleRef={mockPlaceholderExtraStyleRef}
+          />,
+          { createNodeMock: () => mockCanvas },
+        );
+      });
+      const canvas = component.toJSON();
+      expect(canvas.type).toBe('canvas');
+      expect(canvas.children).toBe(null);
+      expect(canvas.props.height).toBe(64);
+      expect(canvas.props.width).toBe(32);
+      // change offset to 0 should not update
+      mockCanvas.offsetWidth = 0;
+      act(() => {
+        mockResizeHandler();
+      });
+      const canvas2 = component.toJSON();
+      expect(canvas2.type).toBe('canvas');
+      expect(canvas2.children).toBe(null);
+      expect(canvas2.props.style.height).toBe(200);
     });
   });