+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 ### Bugfix
 
+- Fix blurhash zero width resizing bug @reebalazs
+- Fix ignoring blurhash width if width (or size) was a string @reebalazs
+- Update fast-blurhash with recent blurhash fixes @reebalazs
+
 ### Internal
 
 ## 12.3.6 (2022-10-20)